#11 - Powerups
The powerups have been MOSTLY implemented. Things left to do:

Make it so sprites change when every powerup is collected, so far only super mushroom does it.

Update the score when powerups are picked up.

Implementation of fireball shooting.

Sound effects and animations of powerup pickups.

EXTRA NOTES:
Star bouncing is a bit buggy because of rigid bodies, will try to fix soon.

I've added 3 new tags and also tagged a few things to test if everything was working, the tags I added are: Wall, Floor and Powerups. I've already tagged all the floors but please tag the walls and floors with the appropriate tag to keep my code working.

I think we should also have a tag for interactable blocks, but that's not implemented yet so I won't do it yet.