#19 Fireball Shooting
Fireballs are fired when clicking the mouse when the Mario is in the Fiery state, fireballs don't do anything yet and are also using the fireworks sprite because I don't know how to do the sprites, so please change it when you can.

Star bouncing was tweaked using bounciness but it will likely need an extra line to get the bounces started when the star gets spawned. I've also tagged anything I consider a wall with "Wall" so the star can bounce off them properly.